Refactor middleware into modular components
