+++ README.md
@@ -9,6 +9,15 @@ AutoDream OS is a modular, V2 standards compliant platform for building agent-dr
 - **Smoke tests and test suites** to validate components
 - **Structured logging and error handling**
 - **CI/CD templates** for Jenkins, GitLab CI and Docker
+- **Refactored middleware pipeline** split into SRP modules under
+  `src/services/middleware`
+
+## Middleware Execution Order
+Middleware chains process packets sequentially in the order that middleware
+components are listed. Each component receives the `DataPacket` returned by the
+previous one, allowing earlier middleware to mutate the packet before later
+components execute. Disabled middleware or components whose conditions fail are
+skipped.
 
 ## Getting Started
 ### Installation
