+++ src/services/middleware/__init__.py
@@ -1,0 +1,20 @@
+"""Middleware package exposing orchestrator and common components."""
+
+from .models import DataPacket, DataFlowDirection, MiddlewareChain, MiddlewareType
+from .base import BaseMiddlewareComponent
+from .orchestrator import MiddlewareOrchestrator
+from .components.routing import RoutingMiddleware
+from .components.transformations import DataTransformationMiddleware
+from .components.validation import ValidationMiddleware
+
+__all__ = [
+    "DataPacket",
+    "DataFlowDirection",
+    "MiddlewareChain",
+    "MiddlewareType",
+    "BaseMiddlewareComponent",
+    "MiddlewareOrchestrator",
+    "RoutingMiddleware",
+    "DataTransformationMiddleware",
+    "ValidationMiddleware",
+]
